+++ apps/extension/src/app/pages/rpc-btc-add-account/register-btc-policy-account.ts
@@ -1,12 +1,5 @@
 import { useCallback } from 'react';
 
-import {
-  compileWshDescriptor,
-  getAddressFromOutScript,
-  getBtcSignerLibNetworkConfigByMode,
-} from '@leather.io/bitcoin';
-import { makeAccountIdentifer } from '@leather.io/crypto';
-import { type NetworkConfiguration } from '@leather.io/models';
 import { type RpcParams, type RpcResult, btcAddAccount } from '@leather.io/rpc';
 
 import { logger } from '@shared/logger';
@@ -19,79 +12,9 @@ import {
   useCurrentNetworkId,
   useNetworks,
 } from '@app/store/networks/networks.selectors';
-import { makePolicyId } from '@app/store/policy/policy-store.utils';
 import { userAddsPolicyAccount } from '@app/store/policy/policy.slice';
 
-interface ResolveBtcPolicyNetworkArgs {
-  paramsNetwork: string | undefined;
-  defaultNetwork: NetworkConfiguration;
-  defaultNetworkId: string;
-  networks: Record<string, NetworkConfiguration>;
-}
-
-function resolveBtcPolicyNetwork({
-  paramsNetwork,
-  defaultNetwork,
-  defaultNetworkId,
-  networks,
-}: ResolveBtcPolicyNetworkArgs) {
-  if (!paramsNetwork) return { network: defaultNetwork, networkId: defaultNetworkId };
-
-  const requestedNetwork = networks[paramsNetwork];
-  if (!requestedNetwork) throw new Error(`Unknown BTC add account network: ${paramsNetwork}`);
-
-  return { network: requestedNetwork, networkId: requestedNetwork.id };
-}
-
-interface CreateBtcPolicyAccountRegistrationArgs {
-  params: RpcParams<typeof btcAddAccount>;
-  fingerprint: string;
-  accountIndex: number;
-  defaultNetwork: NetworkConfiguration;
-  defaultNetworkId: string;
-  networks: Record<string, NetworkConfiguration>;
-}
-
-export function createBtcPolicyAccountRegistration({
-  params,
-  fingerprint,
-  accountIndex,
-  defaultNetwork,
-  defaultNetworkId,
-  networks,
-}: CreateBtcPolicyAccountRegistrationArgs) {
-  const { network, networkId } = resolveBtcPolicyNetwork({
-    paramsNetwork: params.network,
-    defaultNetwork,
-    defaultNetworkId,
-    networks,
-  });
-  const { scriptPubKey } = compileWshDescriptor(params.descriptor);
-  const address = getAddressFromOutScript(
-    scriptPubKey,
-    getBtcSignerLibNetworkConfigByMode(network.chain.bitcoin.mode)
-  );
-  if (!address) throw new Error('Descriptor does not produce an address');
-
-  const parentAccountId = makeAccountIdentifer(fingerprint, accountIndex);
-  const role = 'signer' as const;
-
-  return {
-    addPolicyAccountPayload: {
-      policy: {
-        id: makePolicyId(parentAccountId, address, networkId),
-        parentAccountId,
-        networkId,
-        chain: 'bitcoin' as const,
-        address,
-        descriptor: params.descriptor,
-        role,
-      },
-      name: params.name,
-    },
-    result: { address, descriptor: params.descriptor, accountId: address, role },
-  };
-}
+import { createBtcPolicyAccountRegistration } from './btc-policy-account-registration';
 
 // Derives the policy account's address from the descriptor, saves it to state
 // associated with the active singlesig account, and returns the RPC result.
