+++ apps/extension/src/app/pages/rpc-stx-add-account/register-stx-policy-account.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from 'vitest';
 
 import { type NetworkConfiguration, defaultNetworksKeyedById } from '@leather.io/models';
 
-import { createStxPolicyAccountRegistration } from './register-stx-policy-account';
+import { createStxPolicyAccountRegistration } from './stx-policy-account-registration';
 
 const pubkeyA = '031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f';
 const pubkeyB = '024d4b6cd1361032ca9bd2aeb9d900aa4d45d9ead80ac9423374c451a7254d0766';
